separate tests for file creation
