+++ galata/.env
@@ -1,0 +1,1 @@
+AUTH_URL=https://localhost:8182/ocid1.datasciencenotebooksession.oc1.phx.aaaaaaaarxbdrfby6kbcp4s7wfxzjaildgefltic4mrprffet327tfg3x6mq/
